+++ insanic/testing/helpers.py
@@ -88,15 +88,23 @@ def test_api_endpoint(insanic_application, test_user_token_factory, endpoint, me
     handler = getattr(insanic_application.test_client, method.lower())
 
     request_headers.update({"accept": "application/json"})
+
     if "content-type" not in [h.lower() for h in request_headers.keys()]:
         request_headers.update({"content-type": "application/json"})
-    if "authorization" in [h.lower() for h in request_headers.keys()]:
-        request_headers.update({"authorization": test_user_token_factory(email="test@mmt.com", level=user_level)})
+
+    if "Authorization" in [h.lower() for h in request_headers.keys()]:
+        request_headers.update({"Authorization": test_user_token_factory(email="test@mmt.com", level=user_level)})
+
+    if request_headers.get('content-type', "application/json") == "application/json":
+        handler_kwargs = {"json": request_body}
+    else:
+        handler_kwargs = {"data": request_body}
+
 
     request, response = handler(endpoint,
                                 debug=True,
                                 headers=request_headers,
-                                json=request_body)
+                                **handler_kwargs)
 
     assert expected_response_status == response.status, response.text